pyproject.toml
[project]
name = "crawler-downloader-cli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.10.6",
    "pyyaml>=6.0.2",
    "selenium>=4.25.0",
    "webdriver-manager>=4.0.2",
    "beautifulsoup4>=4.12.3",
    "filetype>=1.2.0",
    "playwright>=1.47.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
